refactor: move verbose flag to core flagutil
