+++ pkg/core/cmdutil/flagutil/debug.go
@@ -1,11 +1,11 @@
 // This file contains functions used to implement the '--debug' command line option.
 
-package debug
+package flagutil
 
 import "github.com/spf13/pflag"
 
-// AddFlag adds the verbose flag to the given set of command line flags.
-func AddFlag(flags *pflag.FlagSet) {
+// VerboseFlag adds the verbose flag to the given set of command line flags.
+func VerboseFlag(flags *pflag.FlagSet) {
 	flags.BoolVarP(
 		&enabled,
 		"verbose",
@@ -15,8 +15,8 @@ func AddFlag(flags *pflag.FlagSet) {
 	)
 }
 
-// Enabled returns a boolean flag that indicates if the verbose mode is enabled
-func Enabled() bool {
+// DebugEnabled returns a boolean flag that indicates if the verbose mode is enabled
+func DebugEnabled() bool {
 	return enabled
 }
